+++ src/app/(main)/(article)/blogEditor/page.jsx
@@ -1,0 +1,10 @@
+import MenuTabs from '@/src/components/header/MenuTabs';
+
+export default function BlogEditor() {
+  return (
+    <div>
+      <MenuTabs />
+      <h1>글쓰기 페이지임</h1>
+    </div>
+  );
+}
