+++ src/app/(main)/(blog)/blogDetail/page.jsx
@@ -1,8 +1,11 @@
+import MenuTabs from '@/src/components/header/MenuTabs';
+
 export default function BlogDetail() {
-    return(
-        <div>
-            <h1>건강한 생활을 위한 10가지 습관</h1>
-            <p>매일 아침 30분 운동과 충분한 수면이 건강의 기본입니다...</p>
-        </div>
-    );
+  return (
+    <div>
+      <MenuTabs />
+      <h1>건강한 생활을 위한 10가지 습관</h1>
+      <p>매일 아침 30분 운동과 충분한 수면이 건강의 기본입니다...</p>
+    </div>
+  );
 }
