+++ src/app/(main)/(blog)/category/page.jsx
@@ -1,9 +1,12 @@
+import MenuTabs from '@/src/components/header/MenuTabs';
+
 const CategoryView = () => {
-    return (
-        <div>
-            <h1>주제별 보기</h1>
-        </div>
-    );
+  return (
+    <div>
+      <MenuTabs />
+      <h1>주제별 보기</h1>
+    </div>
+  );
 };
 
 export default CategoryView;
