+++ src/components/header/MenuTabs.module.css
@@ -1,0 +1,26 @@
+/* MenuTabs.module.css */
+.menuBar {
+    display: flex;
+    border-top: 1px solid #e0e0e0;
+    border-bottom: 1px solid #e0e0e0;
+    font-family: 'NanumGothic', sans-serif;
+    padding: 10px;
+    height: 2em;
+}
+
+.menuLink {
+    font-weight: bold;
+    font-size: 14px;
+    color: #222;
+    margin-right: 32px;
+    border-bottom: 4px solid transparent;
+    text-decoration: none;
+    transition: color 0.2s, border-bottom 0.2s;
+    padding-bottom: 2px;
+}
+
+.menuLink:hover,
+.menuLinkActive {
+    color: #4CAF50;
+    border-bottom: 4px solid #4CAF50;
+}
